+++ src/components/Capture/Face.js
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact'
 import { appendToTracking } from '../../Tracker'
 import Selfie from '../Photo/Selfie'
-import Video from '../Video'
+import FaceVideo from '../FaceVideo'
 import Uploader from '../Uploader'
 import PageTitle from '../PageTitle'
 import withCrossDeviceWhenNoCamera from './withCrossDeviceWhenNoCamera'
@@ -75,23 +75,16 @@ class Face extends Component {
   handleFileSelected = (file) =>
     validateFile(file, this.handleUpload, this.handleError)
 
-  renderUploadFallback = (text, type = 'warning') => {
-    const INPUT_CLASS_MAPPER = {
-      error: 'errorFallbackButton',
-      warning: 'warningFallbackButton',
-    }
-    const fallbackButtonType = INPUT_CLASS_MAPPER[type]
-    return (
-      <CustomFileInput
-        className={theme[fallbackButtonType]}
-        onChange={this.handleFileSelected}
-        accept="image/*"
-        capture="user"
-      >
-        {text}
-      </CustomFileInput>
-    )
-  }
+  renderUploadFallback = (text) => (
+    <CustomFileInput
+      className={theme.errorFallbackButton}
+      onChange={this.handleFileSelected}
+      accept="image/*"
+      capture="user"
+    >
+      {text}
+    </CustomFileInput>
+  )
 
   renderCrossDeviceFallback = (text, callback) => (
     <FallbackButton
@@ -119,7 +112,7 @@ class Face extends Component {
     }
     const cameraProps = {
       renderTitle: <PageTitle title={title} smaller />,
-      containerClassName: style.faceContainer,
+      cameraClassName: style.faceContainer,
       renderFallback: isDesktop
         ? this.renderCrossDeviceFallback
         : this.renderUploadFallback,
@@ -142,10 +135,9 @@ class Face extends Component {
       )
       if (requestedVariant === 'video') {
         return (
-          <Video
+          <FaceVideo
             {...cameraProps}
             onVideoCapture={this.handleVideoCapture}
-            ariaLabel={ariaLabelForSelfieCameraView}
           />
         )
       }