+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 - Internal: Added default filename for all document uploads if filename is not present.
 - Public: Added user consent content reload screen
 - Public: When `photoCaptureFallback` option is disabled in the face step, the requested variant is video and browser does not support MediaRecorder, attempt to redirect to the cross-device flow or display the unsupported browser error.
+- Internal: Refactor for better reusability of video capture components.
 
 ### Changed